fix: cypress trading tests
