+++ apps/trading-e2e/src/integration/wallet-vega.cy.ts
@@ -8,103 +8,115 @@ const manageVegaBtn = 'manage-vega-wallet';
 const form = 'rest-connector-form';
 const dialogContent = 'dialog-content';
 
-describe('connect hosted wallet', { tags: '@smoke' }, () => {
-  beforeEach(() => {
-    // Using portfolio page as it requires vega wallet connection
-    cy.visit('/#/portfolio');
-    cy.mockTradingPage();
-    cy.mockSubscription();
-    cy.get('main[data-testid="/portfolio"]').should('exist');
-  });
-
-  it('can connect', () => {
-    // 0002-WCON-002
-    // 0002-WCON-003
-    // 0002-WCON-039
-    // 0002-WCON-017
-    // 0002-WCON-018
-    // 0002-WCON-019
-
-    // Mock authentication
-    cy.intercept('POST', 'https://wallet.testnet.vega.xyz/api/v1/auth/token', {
-      body: {
-        token: 'test-token',
-      },
+describe(
+  'connect hosted wallet',
+  { tags: '@smoke', testIsolation: true },
+  () => {
+    beforeEach(() => {
+      // Using portfolio page as it requires vega wallet connection
+      cy.visit('/#/portfolio');
+      cy.mockTradingPage();
+      cy.mockSubscription();
+      cy.get('main[data-testid="/portfolio"]').should('exist');
     });
-    // Mock getting keys from wallet
-    cy.intercept('GET', 'https://wallet.testnet.vega.xyz/api/v1/keys', {
-      body: {
-        keys: [
-          {
-            algorithm: {
-              name: 'algo',
-              version: 1,
-            },
-            index: 0,
-            meta: [],
-            pub: 'HOSTED_PUBKEY',
-            tainted: false,
+
+    it('can connect', () => {
+      // 0002-WCON-002
+      // 0002-WCON-003
+      // 0002-WCON-039
+      // 0002-WCON-017
+      // 0002-WCON-018
+      // 0002-WCON-019
+
+      // Mock authentication
+      cy.intercept(
+        'POST',
+        'https://wallet.testnet.vega.xyz/api/v1/auth/token',
+        {
+          body: {
+            token: 'test-token',
           },
-        ],
-      },
+        }
+      );
+      // Mock getting keys from wallet
+      cy.intercept('GET', 'https://wallet.testnet.vega.xyz/api/v1/keys', {
+        body: {
+          keys: [
+            {
+              algorithm: {
+                name: 'algo',
+                version: 1,
+              },
+              index: 0,
+              meta: [],
+              pub: 'HOSTED_PUBKEY',
+              tainted: false,
+            },
+          ],
+        },
+      });
+      cy.getByTestId(connectVegaBtn).click();
+      cy.contains(
+        'Choose wallet app to connect, or to change port or server URL enter a custom wallet location first'
+      );
+      cy.contains('Connect Vega wallet');
+      cy.contains('Hosted Fairground wallet');
+
+      cy.getByTestId('connectors-list')
+        .find('[data-testid="connector-hosted"]')
+        .click();
+      cy.getByTestId(form).find('#wallet').click().type('user');
+      cy.getByTestId(form).find('#passphrase').click().type('pass');
+      cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
+      cy.getByTestId(manageVegaBtn).should('exist');
+      cy.getByTestId('manage-vega-wallet').click();
+      cy.getByTestId('keypair-list').should('exist');
     });
-    cy.getByTestId(connectVegaBtn).click();
-    cy.contains(
-      'Choose wallet app to connect, or to change port or server URL enter a custom wallet location first'
-    );
-    cy.contains('Connect Vega wallet');
-    cy.contains('Hosted Fairground wallet');
 
-    cy.getByTestId('connectors-list')
-      .find('[data-testid="connector-hosted"]')
-      .click();
-    cy.getByTestId(form).find('#wallet').click().type('user');
-    cy.getByTestId(form).find('#passphrase').click().type('pass');
-    cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
-    cy.getByTestId(manageVegaBtn).should('exist');
-    cy.getByTestId('manage-vega-wallet').click();
-    cy.getByTestId('keypair-list').should('exist');
-  });
+    it('doesnt connect with invalid credentials', () => {
+      // 0002-WCON-020
 
-  it('doesnt connect with invalid credentials', () => {
-    // 0002-WCON-020
-
-    // Mock incorrect username/password
-    cy.intercept('POST', 'https://wallet.testnet.vega.xyz/api/v1/auth/token', {
-      body: {
-        error: 'No wallet',
-      },
-      statusCode: 403, // 403 forbidden invalid crednetials
+      // Mock incorrect username/password
+      cy.intercept(
+        'POST',
+        'https://wallet.testnet.vega.xyz/api/v1/auth/token',
+        {
+          body: {
+            error: 'No wallet',
+          },
+          statusCode: 403, // 403 forbidden invalid crednetials
+        }
+      );
+      cy.getByTestId(connectVegaBtn).click();
+      cy.getByTestId('connectors-list')
+        .find('[data-testid="connector-hosted"]')
+        .click();
+      cy.getByTestId(form).find('#wallet').click().type('invalid name');
+      cy.getByTestId(form).find('#passphrase').click().type('invalid password');
+      cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
+      cy.getByTestId('form-error').should('have.text', 'Invalid credentials');
     });
-    cy.getByTestId(connectVegaBtn).click();
-    cy.getByTestId('connectors-list')
-      .find('[data-testid="connector-hosted"]')
-      .click();
-    cy.getByTestId(form).find('#wallet').click().type('invalid name');
-    cy.getByTestId(form).find('#passphrase').click().type('invalid password');
-    cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
-    cy.getByTestId('form-error').should('have.text', 'Invalid credentials');
-  });
 
-  it('doesnt connect with empty fields', () => {
-    cy.getByTestId(connectVegaBtn).click();
-    cy.getByTestId('connectors-list')
-      .find('[data-testid="connector-hosted"]')
-      .click();
+    it('doesnt connect with empty fields', () => {
+      cy.getByTestId(connectVegaBtn).click();
+      cy.getByTestId('connectors-list')
+        .find('[data-testid="connector-hosted"]')
+        .click();
 
-    cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
-    cy.getByTestId(form)
-      .find('#wallet')
-      .next('[data-testid="input-error-text"]')
-      .should('have.text', 'Required');
-    cy.getByTestId(form)
-      .find('#passphrase')
-      .next('[data-testid="input-error-text"]')
-      .should('have.text', 'Required');
-  });
-});
+      cy.getByTestId('rest-connector-form').find('button[type=submit]').click();
+      cy.getByTestId(form)
+        .find('#wallet')
+        .next('[data-testid="input-error-text"]')
+        .should('have.text', 'Required');
+      cy.getByTestId(form)
+        .find('#passphrase')
+        .next('[data-testid="input-error-text"]')
+        .should('have.text', 'Required');
+    });
+  }
+);
 
-describe('connect vega wallet', { tags: '@smoke' }, () => {
+describe('connect vega wallet', { tags: '@smoke', testIsolation: true }, () => {
   beforeEach(() => {
     // Using portfolio page as it requires vega wallet connection
     cy.visit('/#/portfolio');
